解决无法执行命令&加载不到apk文件

+++ src/main/java/com/android/scrcpy/scrcpyfx/HelloController.java
@@ -12,7 +12,6 @@ import javafx.scene.control.*;
 import org.json.JSONObject;
 
 import java.awt.*;
-import java.awt.datatransfer.Clipboard;
 import java.awt.datatransfer.DataFlavor;
 import java.awt.datatransfer.StringSelection;
 import java.awt.datatransfer.Transferable;
@@ -156,12 +155,13 @@ public class HelloController implements Initializable {
                             new Thread(() -> CommandExecution.execCommand("adb shell am start -n " + runMain)).start();
                         } else {
                             updateText("准备安装通信服务 " + apkPath);
-                            CommandExecution.CommandResult commandResult = CommandExecution.execCommand("adb install " + toAndroidPath(apkPath));
-                            updateText("安装服务 " + commandResult);
+                            if (new File(apkPath).exists()) {
+                                CommandExecution.CommandResult commandResult = CommandExecution.execCommand("adb install " + toAndroidPath(apkPath));
+                                if (commandResult.result != 0) updateText("安装服务失败 " + commandResult);
+                                else updateText("安装服务成功 " + commandResult);
+                            } else updateText("APK不存在，请将apk放置到如下位置:" + apkPath);
                         }
-                    } else {
-                        updateText(file.getAbsolutePath() + " 文件不存在");
-                    }
+                    } else updateText(file.getAbsolutePath() + " 文件不存在");
                 }
             } else isSuccess = false;
         }
@@ -206,29 +206,18 @@ public class HelloController implements Initializable {
      * 把文本设置到剪贴板（复制）
      */
     public static void setClipboardString(String text) {
-        // 获取系统剪贴板
-        Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-        // 封装文本内容
-        Transferable trans = new StringSelection(text);
-        // 把文本内容设置到系统剪贴板
-        clipboard.setContents(trans, null);
+        Toolkit.getDefaultToolkit().getSystemClipboard().setContents(new StringSelection(text), null);
     }
 
     /**
      * 从剪贴板中获取文本（粘贴）
      */
     public static String getClipboardString() {
-        // 获取系统剪贴板
-        Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-        // 获取剪贴板中的内容
-        Transferable trans = clipboard.getContents(null);
+        Transferable trans = Toolkit.getDefaultToolkit().getSystemClipboard().getContents(null);
         if (trans != null) {
-            // 判断剪贴板中的内容是否支持文本
             if (trans.isDataFlavorSupported(DataFlavor.stringFlavor)) {
                 try {
-                    // 获取剪贴板中的文本内容
-                    String text = (String) trans.getTransferData(DataFlavor.stringFlavor);
-                    return text;
+                    return (String) trans.getTransferData(DataFlavor.stringFlavor);
                 } catch (Exception e) {
                     //e.printStackTrace();
                 }
